Moved org endpoint to action file so I don't have to worry about the imports. Also commented out search code incase it's effecting the indexing
