+++ ckanext/bcgov/logic/get.py
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
 '''API functions for searching for and getting data from CKAN.'''
-import ckan.plugins.toolkit as toolkit
+# import ckan.plugins.toolkit as toolkit
 
 import logging
 
@@ -148,36 +148,3 @@ def _group_or_org_list(context, data_dict, is_org=False):
         group_list.append(group_detail if all_fields else getattr(group, ref_group_by))
 
     return group_list
-
-@toolkit.side_effect_free
-def organization_list_related(context, data_dict):
-        '''
-        Returns the list of organizations including parent_of and child_of relationships.
-        '''
-        org_list = get_action('organization_list')(context, data_dict)
-
-        # Add the child orgs to the response:
-        for org in org_list:
-            children = []
-            branches = get_organization_branches(org['id'])
-            group_list = model_dictize.group_list_dictize(branches, context)
-            for branch in group_list:
-                d = {}
-                d['title'] = branch['title']
-                children.append(d)
-
-            org['parent_of'] = children
-
-            parents = []
-            branches = get_parent_orgs(org['id'])
-            group_list = model_dictize.group_list_dictize(branches, context)
-            for branch in group_list:
-                d = {}
-                d['title'] = branch['title']
-                parents.append(d)
-            org['child_of'] = parents
-
-        return_dict = {}
-        return_dict['success'] = True
-        return_dict['result'] = org_list
-        return return_dict
